Version: V1_22_1 (14/03/2019)
This is the OpenWSN develop raw with few changes to include the bsp CC2538EM board.
based in the openmote-cc2538.
This version works fine with OpenVisualizer and Coap with security enable or not
but have to work in the multichannel configuration.
In a single channel version (IEEE802154e.c):
ieee154e_vars.singleChannel     = 0;
this version does not work fine.